requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
allure-pytest==2.15.3
    # via -r requirements.in
allure-python-commons==2.15.3
    # via allure-pytest
annotated-types==0.7.0
    # via pydantic
attrs==25.4.0
    # via allure-python-commons
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via pytest
faker==40.1.0
    # via -r requirements.in
greenlet==3.3.0
    # via playwright
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
packaging==25.0
    # via pytest
playwright==1.57.0
    # via
    #   -r requirements.in
    #   pytest-playwright
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements.in
    #   allure-pytest
    #   pytest-asyncio
    #   pytest-base-url
    #   pytest-playwright
pytest-asyncio==1.3.0
    # via -r requirements.in
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-playwright==0.7.2
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-slugify==8.0.4
    # via pytest-playwright
requests==2.32.5
    # via
    #   -r requirements.in
    #   pytest-base-url
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   pyee
    #   pytest-asyncio
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via faker
urllib3==2.6.2
    # via requests
